Make application's year question more friendly
Options before this change:

- freshman
- sophomore
- junior
- senior

Options after this change:

- Below high school
- Nine
- Ten
- Eleven
- Twelve
- College student
- Teacher
- Parent or guardian
- Other

The change from freshman -> nine, sophomore -> ten, and so on makes the
form more friendly to international applicants because the American
terminology for grades is, well... American.

The addition of the other form fields makes the application friendlier
to any applicant that's not in high school (most notably middle
schoolers and teachers).

Closes #201.