Don't rely on ERB templating for Stripe Checkout
Previously we relied on ERB templating to inject the Stripe publishable
key and Hack Club logo URL into our JavaScript to pass to our Stripe
Checkout logic. Unfortunately that doesn't work in production because
the asset pipeline bypasses any templating that might occur.

This change takes advantage of pluggable_js to pass that data in through
a JavaScript object rendered in the view instead.